clean up using IDE tools remove material icons again remove anything that is not related to the export functionality and not yet part of master (i.e. file browser, record stuff)
