+++ EiDiA_frontend/src/components/exportView/ExportMainView.js
@@ -1,17 +1,16 @@
 import React from 'react';
-import DocEditor from "./Subcomponents/DocEditor";
+import DocEditor from "./subcomponents/DocEditor";
 import RightSidepanel from "./RightSidepanel";
-import {RichUtils, EditorState, ContentState} from 'draft-js';
+import {ContentState, convertToRaw, EditorState, RichUtils} from 'draft-js';
 import FloatingWindows from './FloatingWindow';
-import {Row, Column, documentMockData, llorem} from '../../support files/constants';
-import {convertToRaw} from 'draft-js';
+import {Column, documentMockData, llorem, Row} from '../../support files/constants';
 
 function Dialog(props) {
-    if (props.currentPage == "Select Template") {
-        return <div />
+    if (props.currentPage === "Select Template") {
+        return <div/>
     }
     return (
-        <FloatingWindows 
+        <FloatingWindows
             open={props.open}
             onClose={props.onClose}
             save={props.saveTemplate}
@@ -24,8 +23,7 @@ function Dialog(props) {
 
 
 function fetchDocumentData(docList) {
-    const documentData = documentMockData;
-    return documentData;
+    return documentMockData;
 }
 
 function isPath(string) {
@@ -54,7 +52,7 @@ export default class ExportMainView extends React.Component {
     constructor(props) {
         super(props);
         this.editorText = llorem["Template 0"];
-        
+
         this.state = {
             editorState: EditorState.createWithContent(ContentState.createFromText(this.editorText)),
             textAlignment: "left",
@@ -108,10 +106,10 @@ export default class ExportMainView extends React.Component {
 
     // Set Value to variable manually (not from document)
     setValueToVariable(value) {
-        var newState = this.state;
+        let newState = this.state;
 
-        var variableState = newState.variables;
-        let index = variableState[newState.selectedVariable]["index"]; 
+        let variableState = newState.variables;
+        let index = variableState[newState.selectedVariable]["index"];
         variableState[newState.selectedVariable]["value"] = value;
         variableState[newState.selectedVariable]["source"] = value;
         newState.variables = variableState;
@@ -124,7 +122,7 @@ export default class ExportMainView extends React.Component {
 
     // Replace position of that variable with new value
     setValuesToText(index, newValue, editorState) {
-        var editorText = this.getTextFromEditorstate(editorState);
+        let editorText = this.getTextFromEditorstate(editorState);
         const tmp_arr = editorText.split(" ");
         const toReplace = tmp_arr[index];
         editorText = editorText.replace(toReplace, newValue);
@@ -135,14 +133,17 @@ export default class ExportMainView extends React.Component {
     // matches given data from doucment with variables in document text
     mapValues() {
         const documents = fetchDocumentData(this.state.selectedDocs);
-        var newState = this.state;
+        let newState = this.state;
 
-        var variables = this.state.variables;
-        var selectedDocs = this.state.selectedDocs;
-        var documentData = []
-        selectedDocs.forEach((name) => { if (name in documents) documentData.push(documents[name])});
+        let variables = this.state.variables;
+        let selectedDocs = this.state.selectedDocs;
+        let documentData = []
+        selectedDocs.forEach((name) => {
+            if (name in documents) documentData.push(documents[name])
+        });
 
-        var editorText = this.getTextFromEditorstate(newState.editorState);;
+        let editorText = this.getTextFromEditorstate(newState.editorState);
+
         for (let k of Object.keys(variables)) {
             if (isPath(k.slice(1))) {
                 let tmp = k.split("/");
@@ -153,7 +154,7 @@ export default class ExportMainView extends React.Component {
                 let value = documentData[documentIndex][variable];
 
                 // update current value and value source of that variable for state
-                variables[k]["value"] = value;  
+                variables[k]["value"] = value;
                 variables[k]["source"] = "\/" + selectedDocs[documentIndex] + "\/" + variable;
                 editorText = this.setValuesToText(index, value, newState.editorState)
             }
@@ -164,15 +165,15 @@ export default class ExportMainView extends React.Component {
         this.setState(newState);
     }
 
-    
+
     toggleDialog() {
-        var newState = this.state;
+        let newState = this.state;
         newState.open = !newState.open;
         this.setState(newState);
     }
 
     addToList(element) {
-        var newState = this.state;
+        let newState = this.state;
         if (!newState.selectedDocs.includes(element)) {
             newState.selectedDocs.push(element);
             this.setState(newState);
@@ -183,7 +184,7 @@ export default class ExportMainView extends React.Component {
     // collects all variables of the template
     selectTemplate(value) {
         this.editorText = llorem[value] || this.editorText;
-        var newState = this.state;
+        let newState = this.state;
 
         newState.editorState = EditorState.createWithContent(ContentState.createFromText(this.editorText));
         newState.selectedTemplate = value;
@@ -193,7 +194,7 @@ export default class ExportMainView extends React.Component {
 
     toggleInlineStyle(style) {
         const inlineStyle = style.toUpperCase();
-        var newState = this.state;
+        let newState = this.state;
 
         newState.editorState = RichUtils.toggleInlineStyle(this.state.editorState, inlineStyle)
         this.setState(newState);
@@ -201,23 +202,23 @@ export default class ExportMainView extends React.Component {
     }
 
     toggleBlockType(align) {
-        var newState = this.state;
+        let newState = this.state;
 
         newState.editorState = RichUtils.toggleBlockType(this.state.editorState, align);
         newState.textAlignment = align;
         this.setState(newState);
     }
-    
+
 
     // update function for editor when usere give input to the editor
     // scans for new variables entered by user when in "Edit Templatepage"
     onChange(editorState) {
-        var newState = this.state;
+        let newState = this.state;
         newState.editorState = editorState;
-        if (this.props.currentPage == "Edit Template") {
+        if (this.props.currentPage === "Edit Template") {
             newState.variables = this.scanForNewVariables(newState.variables, editorState);
         }
-        
+
         this.setState(newState);
     }
 
@@ -228,7 +229,7 @@ export default class ExportMainView extends React.Component {
         for (let k of Object.keys(newVariables)) {
             let index = (k in currVariables) ? currVariables[k]["index"] : newVariables[k]["index"];
             let value = (k in currVariables) ? currVariables[k]["value"] : newVariables[k]["value"];
-            let source= (k in currVariables) ? currVariables[k]["source"] : newVariables[k]["source"];
+            let source = (k in currVariables) ? currVariables[k]["source"] : newVariables[k]["source"];
             newVariableState[k] = {
                 index: index,
                 value: value,
@@ -241,8 +242,7 @@ export default class ExportMainView extends React.Component {
 
     getTextFromEditorstate(editorState) {
         const blocks = convertToRaw(editorState.getCurrentContent()).blocks;
-        const value = blocks.map(block => (!block.text.trim() && '\n') || block.text).join('\n');
-        return value;
+        return blocks.map(block => (!block.text.trim() && '\n') || block.text).join('\n');
     }
 
     // Collects variables from document text
@@ -250,9 +250,9 @@ export default class ExportMainView extends React.Component {
     extractVariables(editorState) {
         const value = this.getTextFromEditorstate(editorState);
         const arr = value.split(" ");
-        var varObject = {}
+        let varObject = {}
         for (let i = 0; i < arr.length; i++) {
-            if (arr[i][0] == "$") {
+            if (arr[i][0] === "$") {
                 varObject[arr[i]] = {
                     index: i,
                     value: "",
@@ -263,22 +263,22 @@ export default class ExportMainView extends React.Component {
         return varObject;
     }
 
-    componentDidUpdate(prevProps) {
+    componentDidUpdate(prevProps, prevState, snapshot) {
         if (this.props.currentPage !== prevProps.currentPage) {
-            if (this.props.currentPage == "Edit") this.mapValues();
-        }  
+            if (this.props.currentPage === "Edit") this.mapValues();
+        }
     }
-    
+
     //Edit View:  User chose variable to manually assign value
-    setSelectedVariable(event){
-        var newState = this.state;
+    setSelectedVariable(event) {
+        let newState = this.state;
         newState.selectedVariable = event.target.value;
         this.setState(newState);
     }
-    
+
     // TODO: Let User save template
     saveTemplate() {
-        var newState = this.state;
+        let newState = this.state;
         newState.open = false;
         this.setState(newState);
     }
@@ -295,7 +295,7 @@ export default class ExportMainView extends React.Component {
 
     render() {
         const currentPage = this.props.currentPage
-        const editorState=this.state.editorState;
+        const editorState = this.state.editorState;
         const actionSet = this.actionSet[currentPage];
         const componentSet = components[currentPage];
         return (
@@ -304,10 +304,12 @@ export default class ExportMainView extends React.Component {
                     <Column>
                     </Column>
                     <Column>
-                        <DocEditor 
+                        <DocEditor
                             readOnly={this.props.readOnly}
                             textAlignment={this.state.textAlignment}
-                            ref={(docEditor) => {this.docEditor = docEditor}}
+                            ref={(docEditor) => {
+                                this.docEditor = docEditor
+                            }}
                             editorState={editorState}
                             onChange={this.onChange}
                         />
@@ -330,15 +332,15 @@ export default class ExportMainView extends React.Component {
                         />
                     </Column>
                 </Row>
-            <Dialog 
-                open={this.state.open} 
-                onClose={this.toggleDialog} 
-                save={this.saveTemplate} 
-                currentPage={this.props.currentPage}
-                selectedDocs={this.state.selectedDocs}
-                download={this.downloadDocument}
-                editorState={this.state.editorState}    
-            />
+                <Dialog
+                    open={this.state.open}
+                    onClose={this.toggleDialog}
+                    save={this.saveTemplate}
+                    currentPage={this.props.currentPage}
+                    selectedDocs={this.state.selectedDocs}
+                    download={this.downloadDocument}
+                    editorState={this.state.editorState}
+                />
             </div>
         );
     }
