+++ EiDiA_frontend/src/components/exportView/styles.css
@@ -1,41 +1,41 @@
 html {
     height: 100%;
-  }
-  
-  body {
+}
+
+body {
     height: 100%;
     margin: 0;
     background-repeat: no-repeat;
     background-image: linear-gradient(#5b1492, #882973, #ce7eb5);
-  }
-  
-  .btn {
+}
+
+.btn {
     text-align: center;
     position: absolute;
     top: 30%;
     left: 42%;
     zoom: 1.5;
-  }
-  
-  button {
+}
+
+button {
     border-radius: 10px;
     background-color: grey;
     outline: none;
-  }
-  
-  button:hover {
+}
+
+button:hover {
     background-color: white;
-  }
-  
-  .modal {
+}
+
+.modal {
     position: fixed;
     z-index: 1;
     width: 100%;
     height: 100%;
     background-color: rgba(0, 0, 0, 0.25);
-  }
-  
-  .modal_content {
+}
+
+.modal_content {
     background-color: white;
     position: absolute;
     top: 20%;
@@ -44,15 +44,14 @@ html {
     padding: 20px;
     border-radius: 5px;
     border: 2px solid black;
-  }
-  
-  .close {
+}
+
+.close {
     color: Black;
     float: right;
-  }
-  
-  .close:hover {
+}
+
+.close:hover {
     color: cyan;
     cursor: pointer;
-  }
-  
+}
