+++ EiDiA_frontend/src/components/exportView/subcomponents/ExportDocumentWindow.js
@@ -1,21 +1,15 @@
 import React from 'react';
-import { Dialog, DialogTitle, Typography, Button, Box, Input } from '@material-ui/core';
-import PersonIcon from '@material-ui/icons/Person';
-import AddIcon from '@material-ui/icons/Add';
-import {Row, Column} from '../../../support files/constants';
-import {RecordSymbol} from "../../fileCabinet/RecordSymbol";
+import {Button, Input, Typography} from '@material-ui/core';
+import {Column, Row} from '../../../support files/constants';
+import RecordSymbol from "../../fileCabinet/RecordSymbol";
 import styled from "styled-components";
-import SaveTemplateWindow from './SaveTemplateWindow';
-import FormLabel from '@material-ui/core/FormLabel';
-import FormControl from '@material-ui/core/FormControl';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
-import FormHelperText from '@material-ui/core/FormHelperText';
 import Checkbox from '@material-ui/core/Checkbox';
 import Preview from './Preview';
 
 
-const styles={
+const styles = {
     column: {
         margin: "15px",
     },
@@ -34,7 +28,7 @@ const styles={
         margin: "15px",
         align: "right",
         bottom: 0,
-        right:0,
+        right: 0,
     }
 }
 
@@ -51,7 +45,7 @@ export default class ExportDocumentWindow extends React.Component {
     render() {
         let filteredRecords = this.props.filteredRecords;
         let selectedDocs = this.props.selectedDocs;
-        return(
+        return (
             <Row style={{margin: "10px"}}>
                 <Column style={styles.column}>
                     <Typography variant="subtitle2">
@@ -64,7 +58,8 @@ export default class ExportDocumentWindow extends React.Component {
                         Used Documents
                     </Typography>
                     <FormGroup>
-                        {selectedDocs.map((doc) => <FormControlLabel control={<Checkbox color="primary"/>} label={doc}/>)}
+                        {selectedDocs.map((doc) => <FormControlLabel control={<Checkbox color="primary"/>}
+                                                                     label={doc}/>)}
                     </FormGroup>
                 </Column>
                 <Column style={styles.column}>
@@ -73,17 +68,20 @@ export default class ExportDocumentWindow extends React.Component {
                             Template Name
                         </Typography>
                     </Row>
-                    <Row>    
-                    <Input placeholder="Template Name"  inputProps={{ 'aria-label': 'description' }} />
+                    <Row>
+                        <Input placeholder="Template Name" inputProps={{'aria-label': 'description'}}/>
                     </Row>
                     <Row>
                         <div width="80%">
                             <Typography variant="subtitle2">
                                 Template Name
                             </Typography>
-                            <Input placeholder="Search Records ..." fullWidth={true}  inputProps={{ 'aria-label': 'description' }} value={this.props.value} onChange={this.props.updateSearch.bind(this)} />
+                            <Input placeholder="Search Records ..." fullWidth={true}
+                                   inputProps={{'aria-label': 'description'}} value={this.props.value}
+                                   onChange={this.props.updateSearch.bind(this)}/>
                             <FlexRow>
-                                {filteredRecords.map(record => <RecordSymbol fontSize='3vw' labelFontSize='10px' key={record} name={record}/>)}
+                                {filteredRecords.map(record => <RecordSymbol fontSize='3vw' labelFontSize='10px'
+                                                                             key={record} name={record}/>)}
                             </FlexRow>
                         </div>
                     </Row>
