+++ EiDiA_frontend/src/components/exportView/subcomponents/SetValueSection.js
@@ -1,14 +1,11 @@
 import React from 'react';
 import FormControl from '@material-ui/core/FormControl';
-import InputLabel from '@material-ui/core/InputLabel';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import TextField from '@material-ui/core/TextField';
-import {Row, Column} from '../../../support files/constants';
 import FormHelperText from '@material-ui/core/FormHelperText';
 import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
-import AddIcon from '@material-ui/icons/Add';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
@@ -16,7 +13,7 @@ import TableContainer from '@material-ui/core/TableContainer';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
-import {documentMockData} from '../../../support files/constants'; 
+import {MdAdd} from "react-icons/all";
 
 const styles = {
     textField: {
@@ -29,11 +26,11 @@ const styles = {
 }
 
 function createData(variable, value, source) {
-    return { variable, value, source };
-  }
+    return {variable, value, source};
+}
 
-function convertData(variables, value=null) {
-    var out = [];
+function convertData(variables, value = null) {
+    let out = [];
     for (let k of Object.keys(variables)) {
         let variable = variables[k];
         out.push(createData(k, variable["value"], variable["source"]));
@@ -45,7 +42,7 @@ export default class SetValueSection extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            input: "",    
+            input: "",
         }
 
         this.handleChangeTextfield = this.handleChangeTextfield.bind(this);
@@ -65,7 +62,7 @@ export default class SetValueSection extends React.Component {
         const variableState = this.props.variables;
         const variableKeys = Object.keys(variableState);
         const rows = convertData(variableState);
-        return(
+        return (
             <div>
                 <FormControl style={{margin: "5px"}}>
                     <Select variant="outlined" onChange={this.props.onAction2}>
@@ -73,29 +70,31 @@ export default class SetValueSection extends React.Component {
                     </Select>
                     <FormHelperText>Variables</FormHelperText>
                 </FormControl>
-                <TextField style={{margin: "5px"}} variant="outlined" helperText="e.g. 01.01.2020" onChange={this.handleChangeTextfield}/>
-                <IconButton style={{margin: "5px"}} color="primary" onClick={this.setValueToVariable}><AddIcon/></IconButton>
+                <TextField style={{margin: "5px"}} variant="outlined" helperText="e.g. 01.01.2020"
+                           onChange={this.handleChangeTextfield}/>
+                <IconButton style={{margin: "5px"}} color="primary"
+                            onClick={this.setValueToVariable}><MdAdd/></IconButton>
                 <TableContainer component={Paper}>
-                <Table aria-label="simple table">
-                    <TableHead>
-                    <TableRow>
-                        <TableCell>Variable</TableCell>
-                        <TableCell align="right">Value</TableCell>
-                        <TableCell align="right">Source</TableCell>
-                    </TableRow>
-                    </TableHead>
-                    <TableBody>
-                    {rows.map((row) => (
-                        <TableRow key={row.name}>
-                        <TableCell component="th" scope="row">
-                            {row.variable}
-                        </TableCell>
-                        <TableCell align="right">{row.value}</TableCell>
-                        <TableCell align="right">{row.source}</TableCell>
-                        </TableRow>
-                    ))}
-                    </TableBody>
-                </Table>
+                    <Table aria-label="simple table">
+                        <TableHead>
+                            <TableRow>
+                                <TableCell>Variable</TableCell>
+                                <TableCell align="right">Value</TableCell>
+                                <TableCell align="right">Source</TableCell>
+                            </TableRow>
+                        </TableHead>
+                        <TableBody>
+                            {rows.map((row) => (
+                                <TableRow key={row.name}>
+                                    <TableCell component="th" scope="row">
+                                        {row.variable}
+                                    </TableCell>
+                                    <TableCell align="right">{row.value}</TableCell>
+                                    <TableCell align="right">{row.source}</TableCell>
+                                </TableRow>
+                            ))}
+                        </TableBody>
+                    </Table>
                 </TableContainer>
                 <div>
                     <Button
