+++ EiDiA_frontend/src/components/exportView/subcomponents/TemplateList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { List, Typography } from '@material-ui/core';
+import {List, Typography} from '@material-ui/core';
 import TemplateListItem from './TemplateListItem';
 
 export default class TemplateList extends React.Component {
@@ -18,13 +18,13 @@ export default class TemplateList extends React.Component {
     }
 
     fetchTemplates() {
-        var newState = this.state;
+        let newState = this.state;
         newState.templateList = ["Template 0", "Template 1", "Template 2"];
         this.setState(newState);
     }
 
     handleListItemclick(value, index) {
-        var newState = this.state;
+        let newState = this.state;
         newState.selectedIndex = index;
         this.setState(newState);
         this.props.onAction1(value);
@@ -32,26 +32,25 @@ export default class TemplateList extends React.Component {
 
     render() {
         const items = this.state.templateList;
-        return(
+        return (
             <div>
                 <Typography variant="subtitle2">
                     Templates
                 </Typography>
                 <List dense={true}>
-                    {items.map((item, index) => 
-                        <TemplateListItem 
+                    {items.map((item, index) =>
+                        <TemplateListItem
                             key={item}
                             text={item}
                             id={index}
                             index={index}
                             isSelected={this.state.selectedIndex === index}
                             onAction={this.handleListItemclick}
-                    />
+                        />
                     )}
-                    
+
                 </List>
             </div>
         );
     }
-
 }
