+++ EiDiA_frontend/src/components/exportView/subcomponents/VariableList.js
@@ -3,14 +3,14 @@ import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
 
 export default class VariableList extends React.Component {
-    constructor(props){
+    constructor(props) {
         super(props);
     }
 
-    render(){
+    render() {
         const variableState = this.props.variables;
         const variableKeys = Object.keys(variableState);
-        return(
+        return (
             <div style={{margin: "3%"}}>
                 <p/>
                 Use <input defaultValue="3" width="10px" type="number"/> documents.
