+++ package.json
@@ -44,9 +44,8 @@
     "webfontloader": "^1.6.28"
   },
   "devDependencies": {
-    "@material-ui/icons": "^4.9.1",
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.5",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-plugin-transform-runtime": "^6.23.0",
